mcp_langchainbot/requirements.txt
langchain==0.3.25
langchain-mcp-adapters==0.1.7
mcp==1.10.1
python-dotenv==1.1.0
langchain-deepseek==0.1.3
tqdm==4.67.1